# apigateway/internal/docs/swagger.yaml
basePath: /
definitions:
  models.CreateItemRequest:
    properties:
      amount:
        type: integer
      name:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
  models.Error:
    properties:
      message:
        type: string
    type: object
  models.GeneralItem:
    properties:
      amount:
        type: integer
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      username:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is an API for managing items.
  title: ITEMS API
  version: "2.0"
paths:
  /items:
    get:
      consumes:
      - application/json
      description: Retrieves all items.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GeneralItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get all items
      tags:
      - items
  /items/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing item by its ID and broadcasts the delete event
        via Kafka.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete an item by ID
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Retrieves an item by its ID.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GeneralItem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get an item by ID
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Updates an existing item by its ID and broadcasts the update event
        via Kafka.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item to update
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.GeneralItem'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update an item by ID
      tags:
      - items
  /items/create:
    post:
      consumes:
      - application/json
      description: Creates a new item and broadcasts the creation event via Kafka.
      parameters:
      - description: Item to create
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.CreateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create a new item
      tags:
      - items
schemes:
- https
swagger: "2.0"